requirements.txt
discord.py
pymongo
PyDrive
google-api-python-client
google-auth-httplib2
google-auth-oauthlib